Adds the text 'Latest' on the homepage to confirm update on Ubuntu VM.

+++ app/Http/Controllers/AdminsController.php
@@ -13,21 +13,16 @@ use App\Http\Requests;
 
 class AdminsController extends Controller {
 
-    public function confirmAdmin() {
-      $user = \Auth::user();
-      if($user->isAdmin) {
-        return redirect()->action('AdminsController@manageUsers');
-      }
-    }
-
-
     public function __construct() {
       $this->middleware('admin');
     }
 
-    public function manageUsers() {
-      $users = User::get();
-      return view('users.index', compact('users'));
+    public function confirmAdmin() {
+      $user = \Auth::user();
+      if($user->isAdmin) {
+        $users = User::get();
+        return view('users.index', compact('users'));
+      }
     }
 
     public function show(User $user) {
@@ -41,8 +36,12 @@ class AdminsController extends Controller {
     }
 
     public function delete(User $user) {
+      if ($user->isAdmin == 0) {
         $user->delete();
-        return redirect('/users/manageUsers');
+        return redirect('/admin');
+      } else {
+        return view('warnings.notOwner');
+      }
     }
 
     public function setAdminWarning(User $user) {
@@ -56,7 +55,7 @@ class AdminsController extends Controller {
     }
 
     public function home() {
-      return redirect('/users/manageUsers');
+      return redirect('/admin');
     }
 
 }
